fix(frontend): Corrige erros de build do Next.js
